Access `add_subparsers()` kwargs
Specifically the ability to pass in things like the `metavar` keyword. It would be nice to just pass `metavar='COMMAND'` to `add_subparsers()` rather than getting using a list of all commands as a metavar. This looks really ugly when you have 10+ commands, and it's straining to fit them all into the `{,}` list.